enhance Thunder identity provider management with new routes and components and update the add-environment-thunder.sh
